chore(vscode-bazel-bsp): checked-in copy of bsp types
Summary:
We have a pending pull request to add BSP TypeScript codegen:
https://github.com/agluszak/bsp-generators/pull/24

Once this is merged, it can be published as a package and will be imported.  In the meantime, checking in this copy of a current version of the generated code, which will allow us to begin working on the client implementation in parallel.