feat(vscode-bazel-bsp): navigate to build files + .bazelproject file from test explorer
Summary:
Adds file navigation from the test explorer.  The root test item will navigate to the .bazelproject file, and test targets will go to the corresponding build file for that target.

To support this, moved the bazel project location into a setting, which gets used both during install and in determining the URI for these test items.  This will also be relevant to trigger a server re-install prompt when the user changes a setting that might require a re-run of the install script.