feat(vscode-bazel-bsp): handle coverage reports via new build/publishOutput notification
Summary:
In finalizing the bazel-bsp server PR to support code coverage (https://github.com/JetBrains/bazel-bsp/pull/563), we ended up going with a separate build/publishOutput notification to send the path to the coverage report.

This update adds handling for this new notification type (in the bsp-ext file for now until it's fully included in the protocol), instead of getting the code coverage report via a task progress notifciation.