fix(vscode-bazel-bsp): avoid duplicate test controller name in tests
Summary:
Tests may be running in a VS Code instance where the actual Bazel BSP extension is installed.  Because the test controller name must be globally unique, this creates errors if using the real test controller ID in tests, as it'll cause a conflict if it's already taken by the real one.  This has newly surfaced because in the next diff, I have a test case that opens an actual Java test file, which is one of the activation events for the extension.

This update pulls the test controller out into its own top level provider, when can then be injected into the TestCaseStore.  In tests, it is then replaced with a mock version with an ID that will not conflict with the main one.