feat(vscode-bazel-bsp): Auto-install BSP server based on setting
Summary:
This update adds a setting that controls the BSP server download-install behavior.

If no existing installation is found in a given repo, we now offer the option to configure behavior based on settings.  In our setup, we will use the "Auto" value, which will allow the server to install in the background without user intervention.