fix(vscode-bazel-bsp): type error while initiating some runs
Summary:
Fixing a runtime error that can occur when attempting to look up an item's lookup key when the item's test case details haven't been set.  mapTestCaseInfoToLookupKey will be able to return undefined instead of throwing an uncaught exception.

Also fix a similar potential issue in prepareTestRunParams (and clean up a duplicate version of prepareTestRunParams in TestItemTestCaseInfo because it duplicates the behavior its the super class).