feat(vscode-bazel-bsp): add full test output to the top level test item in a run
Summary:
At the end of a test run, the response will include a collector with the full contents of stdout during the run.

This output will be added to the top level item in the run.  This is in addition to any individual failures reported from the testFinish notifications during a run.

Also making a tweak to the formatting of the messages reported on individual test cases.  It seems that 'null' values are getting reported by the server as strings, which causes unexpected formatting in some cases.  Treat these 'null' strings as undefined for now while I work on an upstream PR to fix this on the server side.