include minimum jvm version in launch command (#10)
(migrating existing commit from internal repo)

Specify minimum --jvm argument when launching the BSP install command.
This will allow coursier to attempt to find a version that matches if
already installed, or download one that meets the requirements. This
ensures a successful launch even in environments that do not have a new
enough jvm version installed.