Adjustments to Python test discovery (#17)
Making a few tweaks to improve Python test case discovery in files,
based on some initial user feedback:
- Use selection range for positioning of arrows, so arrow lines up with
the function header. The previous range value began at the start of any
decorators or other content above the header.
  Example of adjusted positions:

![image](https://github.com/uber/vscode-bazel-bsp/assets/92764374/ec3096f8-4ab3-420b-9f18-20cccc8551f9)

- Let class names start _or_ end with Test
- Add a filter before returning that excludes classes containing no test
cases. This will reduce noise in cases where the word Test is in a class
name but it is not in fact a test.