Resolve based on open files (#24)
Adds support to discover targets based on a user's open files, allowing
test cases in a file to be resolved even if the user hasn't specifically
expanded the target in the Test Explorer tree.

During the sync process, we will now check a user's open files. If any
are determined to be test files, we request the file's target from the
BSP server, then resolve the test cases in that target. After this is
complete, the user will be able to run tests in those files via run
arrows, without first having to navigate to the test explorer tree. This
is opt out via settings, as it does result in additional back and forth
with the BSP server.

As a follow-up, I plan to add an inlay hint or other info to help the
user update their scope if we try to find the target.