Add debug support in Test Explorer (#29)
Add debug support to the test explorer.

When this feature is enabled (via `bazelbsp.debug.enabled` setting), an
additional debug run option will appear on each test item.

Overall flow:
- When a debug button for a test case is clicked, run is trigged as
usual, with added `additionalBazelParams` (now supported by bazel-bsp
server). These flags are configurable based on
`bazelbsp.debug.bazelFlags` and should include any flags needed to get
Bazel to build and launch the target in debug mode.
- As the build runs, messages are watched for a debug ready message. The
pattern to watch for is configurable in `bazelbsp.debug.readyPattern`.
- Finally, once the pattern is matched, the VS Code launch configuration
is triggered. This allows use of an existing launch configuration in the
workspace (which can be set via settings, workspace, launch.json, etc),
configurable via `bazelbsp.debug.launchConfigName`.

If any of the settings are not properly set, the run will still attempt
to proceed but print a warning in the output.