bazel-bsp server auto-update based on configured version (#30)
Add a version check during the server initialization process. When the
installed version does not match what is configured in VS Code settings
(`bazelbsp.serverVersion`), the installer will be run to get the
required version. This will ensure that a user's installed bazel-bsp
version stays in sync with the version expected by this extension.

Also update jdk version used by Coursier as newer versions of bazel-bsp
require Java 17+, to get this working. As a follow-up we can add in some
additional logic here to depend on system Java where its version is new
enough.