Update bazel-bsp default version (#32)
Updates the default bazel-bsp server version that gets installed by this
extension. There is also a new feature flags data field in the
initialization request that needs to be set here (currently only Python
is behind a flag, Java is always enabled).

This version is necessary to support newly added debug support via the
test explorer. With the newly added changes in
https://github.com/uber/vscode-bazel-bsp/pull/30, users that have a
different version than the one specified will be auto-updated to the
matching version.

Verified core extension workflows with this version:
- Syncing test targets
- Run with/without coverage
- Run with debug