Use namespaced package name in package.json (#36)
For improved security, include a namespace in the package name. This
helps address a potential vulnerability if a user were to attempt to
install an npm package with this same name, which to date we have not
been using.

Since the vsce tool does not support namespaces in the package name, we
also have to update the package command to temporarily remove then
replace the namespace during this step.