Handle undefined result from executeDocumentSymbolProvider (#39)
When resolving test symbols in a document, it's possible that there are
situations where the call to get document symbols will return undefined
(mainly if language server is still activating). This was resulting in
an uncaught exception that causes the loading to appear "stuck" on the
UI. This adds correct handling if this result is undefined.