Patch the Python aspect template after server install (#41)
The Python aspect template installed by the BSP server will only add a
load statement for rules_python if the Bazel version is 8+. In our
setup, we need this loaded always.

We have added this upstream PR
(https://github.com/JetBrains/hirschgarten/pull/210), and also currently
have some internal logic that patches this for our users as a
workaround. However, I think the best place for it would be to add it
here for now so it always runs immediately after install. We can remove
this once we resume our own internal server releases.