Added coursier installation support for mac based environments (#38)
OpenJDK java does not appear on mac based system
https://get-coursier.io/docs/cli-java#managed-jvms in managed jvm.
Also the installation URL are different for Mac
https://get-coursier.io/docs/cli-installation#native-launcher

Here depending on the OS architecture we decide which coursier url and
java version to pick.

For test we use `proxyquire` as dependency to mock os environment.

Updated yarn lock file and changed the rule prefixes from `node/` to
`n/` to match the current naming convention of the eslint-plugin-n
package (formerly eslint-plugin-node)


Test:

Tested the installation on mac and linux environments
Also checked test runs there