Restore previously loaded test cases (#52)
This change will persist the prior list of test cases as part of the
workspace state, and restore this data when the test explorer
relaunches. This avoids a wait for the full BSP build target request
before test cases become visible:
- Update the test case store to cache the build targets result and each
target sources result. When a refresh runs, the latest version of these
get stored.
- Before registering the handlers, there is now an additional step to
pull these stored test cases and re-add them back to the test explorer.
- These get processed through the usual test case creation logic, so
that a test explorer tree created from a restored result or a new one
from the build server will behave the same way.
- The refresh run always clears out any cached values, so these are only
used when restoring the window, and an actual refresh run is always
guaranteed to contain the latest results from the build server.