Add ideClient tag via additional bazel params (#56)
Here we add detectIdeClient utility function to identify which IDEClient
the bsp client is running on.
By default we support cursor and vscode only for now but this can be
modified in future to accomodate more clients.

Here we are using bazel's [test env
flag](https://bazel.build/reference/command-line-reference#flag--test_env)
`--test_env=<ide>` to add ide client info.

Added test to verify the behaviour.

Coverage does not support additionalBazelParams once its supported we
will enable this for coverage as well

IDE is identified by bundleidentifier on mac or path env variable which
is when.

Added gitignore to ignore vsix output.